WIP. Switching to controller tests.
