+++ Manipulator/ControllerManipulator.php
@@ -243,21 +243,6 @@ class ControllerManipulator extends AbstractManipulator
     }
 
     /**
-     * @param ClassMetadata $metadata
-     * @return mixed
-     */
-    protected function getEntityIdentifier(ClassMetadata $metadata)
-    {
-        if (count($metadata->identifier) !== 1) {
-            throw new \InvalidArgumentException(
-                'TdnPilotBundle is incompatible with entities that contain more than one identifier or no identifier.'
-            );
-        }
-
-        return array_values($metadata->identifier)[0];
-    }
-
-    /**
      * @param string $idType
      * @return string
      */
