+++ Resources/skeleton/controller/actions/delete.php.twig
@@ -26,7 +26,7 @@
      *
 {% endif %}
 {% endblock documentation %}
-     * @View(statusCode=204)
+     * @Rest\View(statusCode=204)
      *
      * @param $id
      * @internal {{ entity }}Interface ${{ entity|lowerfirst }}
