+++ Resources/skeleton/controller/actions/getAll.php.twig
@@ -43,7 +43,7 @@
      *   description="Filter by fields. Must be an array ie. &filters[id]=3"
      * )
      *
-     * @View(serializerEnableMaxDepthChecks=true)
+     * @Rest\View(serializerEnableMaxDepthChecks=true)
      *
      * @param ParamFetcherInterface $paramFetcher
      *
