+++ Resources/skeleton/controller/actions/getById.php.twig
@@ -26,7 +26,7 @@
      *
 {% endif %}
 {% endblock documentation %}
-     * @View(serializerEnableMaxDepthChecks=true)
+     * @Rest\View(serializerEnableMaxDepthChecks=true)
      *
      * @param $id
      *
