+++ Resources/skeleton/controller/actions/patch.php.twig
@@ -28,7 +28,7 @@
      *
 {% endif %}
 {% endblock documentation %}
-     * @View(serializerEnableMaxDepthChecks=true)
+     * @Rest\View(serializerEnableMaxDepthChecks=true)
      *
      * @param Request $request
      * @param $id
