+++ Resources/skeleton/controller/actions/put.php.twig
@@ -21,7 +21,7 @@
      *
 {% endif %}
 {% endblock documentation %}
-     * @View(serializerEnableMaxDepthChecks=true)
+     * @Rest\View(serializerEnableMaxDepthChecks=true)
      *
      * @param Request $request
      * @param $id
